--- tests/mcp/cloudrun/locals.tf
@@ -1,1 +1,0 @@
-../../../mcp/cloudrun_locals.tf
