+++ tests/mcp/deployment/GCP/terraform/main.tf
