+++ tests/mcp/deployment/k8s/terraform/main.tf
