--- tests/mcp/gae/manifest_construction/gae_locals.tf
@@ -1,1 +1,0 @@
-../../../../mcp/gae_locals.tf
