--- tests/mcp/kubernetes/terraform/kubernetes_locals.tf
@@ -1,1 +1,0 @@
-../../../../mcp/kubernetes_locals.tf
