--- tests/mcp/kubernetes/terraform/kubernetes_vars.tf
@@ -1,1 +1,0 @@
-../../../../mcp/kubernetes_vars.tf
