+++ tests/mcp/unit_tests/cloudrun/cloudrun_locals.tf
@@ -1,0 +1,1 @@
+../../../../mcp/cloudrun_locals.tf
