+++ tests/mcp/unit_tests/gae/manifest_construction/gae_locals.tf
@@ -1,0 +1,1 @@
+../../../../../mcp/gae_locals.tf
