+++ tests/mcp/unit_tests/gae/manifest_construction/resources/k8s_deployment.yaml
