+++ tests/mcp/unit_tests/gae/manifest_construction/resources/project.yml
