+++ tests/mcp/unit_tests/gae/manifest_construction/resources/single_manifest.tfvars
