+++ tests/mcp/unit_tests/kubernetes/resources/binary.bin
