+++ tests/mcp/unit_tests/kubernetes/resources/mosquitto.conf
