+++ tests/mcp/unit_tests/kubernetes/resources/secret.file
