+++ tests/mcp/unit_tests/kubernetes/terraform/kubernetes_locals.tf
@@ -1,0 +1,1 @@
+../../../../../mcp/kubernetes_locals.tf
