+++ tests/mcp/unit_tests/kubernetes/terraform/kubernetes_vars.tf
@@ -1,0 +1,1 @@
+../../../../../mcp/kubernetes_vars.tf
