+++ tests/mcp/unit_tests/kubernetes/terraform/main.tf
