+++ tests/mcp/unit_tests/kubernetes/terraform/test_deployment.py
@@ -4,18 +4,17 @@ Terraform external provider just handles strings in maps, so tests need to consi
 from sys import path, stderr
 
 try:
-    path.insert(1, '../../../test_fixtures/python_validator')
+    path.insert(1, '../../../../test_fixtures/python_validator')
     from python_validator import python_validator
 except Exception as e:
     print(e, file=stderr)
 
 
 @python_validator
-def test_ingress(query):
+def test_deployment(query):
 
     expected_data = {
-        "service_name" : "service",
-        "service_port" : "8080"
+        "app":"mosquitto"
     }
 
     if query == expected_data:
@@ -27,4 +26,4 @@ def test_ingress(query):
 
 
 if __name__ == '__main__':
-    test_ingress()
+    test_deployment()
