+++ tests/mcp/unit_tests/kubernetes/terraform/test_ingress.py
@@ -4,18 +4,18 @@ Terraform external provider just handles strings in maps, so tests need to consi
 from sys import path, stderr
 
 try:
-    path.insert(1, '../../../test_fixtures/python_validator')
+    path.insert(1, '../../../../test_fixtures/python_validator')
     from python_validator import python_validator
 except Exception as e:
     print(e, file=stderr)
 
 
 @python_validator
-def test_pod(query):
+def test_ingress(query):
 
     expected_data = {
-        "name" : "environment",
-        "value" : "test"
+        "service_name" : "service",
+        "service_port" : "8080"
     }
 
     if query == expected_data:
@@ -27,4 +27,4 @@ def test_pod(query):
 
 
 if __name__ == '__main__':
-    test_pod()
+    test_ingress()
