+++ tests/mcp/unit_tests/kubernetes/terraform/test_pod.py
@@ -1,0 +1,30 @@
+"""
+Terraform external provider just handles strings in maps, so tests need to consider this
+"""
+from sys import path, stderr
+
+try:
+    path.insert(1, '../../../../test_fixtures/python_validator')
+    from python_validator import python_validator
+except Exception as e:
+    print(e, file=stderr)
+
+
+@python_validator
+def test_pod(query):
+
+    expected_data = {
+        "name" : "environment",
+        "value" : "test"
+    }
+
+    if query == expected_data:
+        return {"result": "pass"}
+    else:
+        return {"result" : "fail",
+                "expected" : "{}".format(expected_data),
+                "received" : "{}".format(query)}
+
+
+if __name__ == '__main__':
+    test_pod()
