+++ tests/mcp/unit_tests/kubernetes/terraform/test_pod_autoscaler.py
@@ -4,17 +4,18 @@ Terraform external provider just handles strings in maps, so tests need to consi
 from sys import path, stderr
 
 try:
-    path.insert(1, '../../../test_fixtures/python_validator')
+    path.insert(1, '../../../../test_fixtures/python_validator')
     from python_validator import python_validator
 except Exception as e:
     print(e, file=stderr)
 
 
 @python_validator
-def test_deployment(query):
+def test_pod_autoscaler(query):
 
     expected_data = {
-        "app":"mosquitto"
+        "kind" : "Deployment",
+        "name" : "MyApp"
     }
 
     if query == expected_data:
@@ -26,4 +27,4 @@ def test_deployment(query):
 
 
 if __name__ == '__main__':
-    test_deployment()
+    test_pod_autoscaler()
