+++ tests/mcp/unit_tests/kubernetes/terraform/test_secret.py
@@ -4,17 +4,20 @@ Terraform external provider just handles strings in maps, so tests need to consi
 from sys import path, stderr
 
 try:
-    path.insert(1, '../../../test_fixtures/python_validator')
+    path.insert(1, '../../../../test_fixtures/python_validator')
     from python_validator import python_validator
 except Exception as e:
     print(e, file=stderr)
 
 
 @python_validator
-def test_service(query):
+def test_secret(query):
 
     expected_data = {
-        "app" : "mosquitto",
+
+        "login": "login",
+        "password": "password"
+
     }
 
     if query == expected_data:
@@ -26,4 +29,4 @@ def test_service(query):
 
 
 if __name__ == '__main__':
-    test_service()
+    test_secret()
