+++ tests/mcp/unit_tests/kubernetes/terraform/test_service.py
@@ -4,18 +4,17 @@ Terraform external provider just handles strings in maps, so tests need to consi
 from sys import path, stderr
 
 try:
-    path.insert(1, '../../../test_fixtures/python_validator')
+    path.insert(1, '../../../../test_fixtures/python_validator')
     from python_validator import python_validator
 except Exception as e:
     print(e, file=stderr)
 
 
 @python_validator
-def test_pod_autoscaler(query):
+def test_service(query):
 
     expected_data = {
-        "kind" : "Deployment",
-        "name" : "MyApp"
+        "app" : "mosquitto",
     }
 
     if query == expected_data:
@@ -27,4 +26,4 @@ def test_pod_autoscaler(query):
 
 
 if __name__ == '__main__':
-    test_pod_autoscaler()
+    test_service()
